+++ src/core/h_compiler_utils.py
@@ -70,7 +70,7 @@ class Struct(object):
     def get_type_name(self):
         return "__hascal__" + self.name + self.ptr_str
     
-    def get_type_name_for_error(self):
+    def get_name_for_error(self):
         return self.name + self.ptr_str.replace("*","^")
 
 class Enum(Struct):